+++ changelog.md
@@ -5,6 +5,31 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2017-12-17
+
+### Added
+
+- settings page for setting default value of days to postpone
+- screenshot for settings page
+
+### Changed
+
+- formatting error messages correctly
+- specific error messages for invalid values
+- updated screenshots
+
+### Deprecated
+
+### Removed
+
+### Fixed
+
+- link to readme.txt in readme.md
+- typos and layout of readme.txt
+- minor bug in input field for days (max/min value and id)
+
+### Security
+
 ## [0.1.0] - 2017-12-10
 
 ### Added
@@ -14,4 +39,3 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - readmes
 - changelogs
 - development files
-